refactor: isolate git diff parsing boundaries
